Use the "Å" button to produce "/"
This should make it easier to type the "/" character.

I had problems with this one. After changing it and changing to this
layout, the key still produced the "å" character. I thought that I had
defined it in the wrong way. But this is a valid way to define it. It
turns out that deleting a cache solved the problem[1].

I don't remember why I chose the other things, like the "~" for the second
level. Either it was because I wanted a more convenient way to write the
"~" character---I use dead keys, which makes this character inconvenient
to type---or they were just placeholders carried over from copy pasting.

[1] http://askubuntu.com/a/334100/136104