Add "|" and "§" to third and fourth level of key
I might end up using the key that I use to type the "|" for something
else, so I need a replacement. I don't really use "§", but I can replace
that with something else later.

Even though I only have two levels defined for this key, it seems that,
in the absence of any new definitions, the old symbols are carried over.
That's my theory. The reason I think that is that I right now (not
having recompiled the configuration) have dead keys on the third and
fourth levels.