Add U+2218 RING OPERATOR to fourth level
It seems that keys that I don't redefine in this "nodeadkeys" section
won't simply become "unmapped", but instead use the key defined in the
"basic" section. (See the previous commit message, where I also iscussed
this behaviour.) So in this cases, if I don't remap the fourth level, I
will get the fourth level for "basic" which is "×". And I've defined
this character for the third level, so that is not useful to me.

I guess this makes sense, since the line

    include "no_pr(basic)"

means to include all configurations in "basic", and only override the
things that are explicitly overridden. And if I leave out some levels,
then the keys defined in "basic" for those level will have effect.