Move over code for keyboard setup dotfiles repo
I'm creating a dedicated script for my keyboard setup. This is because I
think this is important enough to have its own script and repository.

The current code sets up my layout and options using setxkbmap. Then
the "Return" key is dual-purposed to be a modifier key when pressed
down and chorded with other keys. This is achieved using xmodmap and
xcape.

See also
========

I also have a repository for my two custom xkb configuration files,
"no_p1" and "no_1". Those are the layouts that gets chosen in the
`setxkbmap` call.