Make the <LSGT> key a level 5 shift key
I have never liked using this key to make the "less than" and "greater
than" signs, and I have set up two other keys to make them more
comfortably. I want to use this key as a level 5 shift.

I intend to make a level 5 shift on the right side of the keyboard as
well, so that I can touch type like I do with the regular shift keys.
Right now I intend to make <Return> that key.

Note that this option also makes it so that level 5 is locked when this
key is pressed when pressed together with another level 5 chooser. I
don't know if I want that. But I think that this is unlikely to happen
by accident, at the least.