Remove all characters from "BKSL" key
To signify that it is not intended to work as a key for producing
characters; I am dual-using it as a "compose" key, and as a level 3
shift.

I don't know if this has any effect, considering how xkb uses
inheritance; maybe this just means that the key will inherit all
mappings from "latin(type2)", since it includes that configuration?

Note: Using "NoSymbol" won't work. I've tried it, and it lead to some
weird behaviour when using this key as a shift level 3 key. It didn't
function as a shift level 3 key anymore, but instead as a shift level 2
key. Except it also cancelled out the "Shift_R" key if it was pressed at
the same time, in the sense that if they were pressed at the same time
done would get small alphanumeric letters, not uppercase ones. This is
weird since when I try to press the left and right shifts at the same
time, I can't write any alphanumeric letters. So then then the "BKSL"
apparently didn't function as one of the shift keys, since it didn't
have identical behaviour to them. I also think this lead to the
other (normally) shift level 3 key to act the same way (the "Caps_Lock"
key).