Don't use two different key codes
For the "Return" and "Multi_key" keys.

Setting a single new, previously unused keycode for each of these keys
suffices, since when the keys are used as "Return" and "Multi_key" they
have the same keycode, anyway. For example, both the "LSGT" and "BKSL"
get recognised as "Multi_key" through the same keycode when they are
pressed and released quickly.