Map "LSGT" and "BKSL" to be shift keys
I assign "BKSL" to be a level 5 shift key and "LSGT" to be a
dual-purpose "compose" key and level 3 shift key in a script that my
Xmonad startup script calls. But it seems that these keys are still
treated as shiftable keys when a level 2 shift key is pressed down and
held before they are pressed. So maybe this will lead them to not be
regarded as shiftable keys any more.

I removed a line in the "nodeadkeys" configuration so as to not risk the
"BKSL" becoming a shiftable key again.