Fix issues with the "BKSL" and "LSGT" keys
First of all I had to change:

    modifier_map ISO_Level3_Shift,

to:

    modifier_map mod5

Since the first one was invalid.

But this was not enough, since "BKSL" still got registered as keysym
"NoSymbol" when pressed after "Shift_R". How I solved this was to map
all eight levels of "BKSL" and "LSGT" to their respective shift levels.

How I found out that that was a way to solve my problem, was to look at
one of the files in the xkb directory in which some "shiftable" (i.e.,
normally used as an alphanumeric or some other "character key) key is
used as some kind of shift key. I found this file:

    /usr/share/X11/xkb/symbols/lv

Which contains this line:

    key <AC11> {[ISO_Level3_Latch, quotedbl, apostrophe, quotedbl]};

From this I figured that I could solve my problem by mapping
all *levels* of the key to the relevant shift function. And apparently
that was all it took.

Note: The order of the `modifier_map` lines and the declaration of the
key levels might also be significant.