Don't use "Alt_R" as level 3 shift
I am using "Caps_Lock" and BKSL as level 3 shift keys, so I don't need
this key to act as a level shift 3 key any more.

This change makes it so that the key falls back on functioning as
another "Alt" key or "Meta" key, i.e. the same as the "Alt_L" button.
This means that I have a symmetrical "Alt" setup, just like I do for
"Ctrl", "Win", and the shift level selectors. And this in turn means
that I can touch type with the "Alt" keys.