Add "TLDE" as an additional "compose" key
I need a compose key for the left side of the keyboard as well.

I have already set up "Caps Lock" and "BKSL" as dual-purpose compose and
shift keys, but the downside of using these as compose keys is that
there is inevitable delay because they are used as two different keys.
Therefore, if I press one of these two as the compose key and one of the
compose sequences in rapid succession, it doesn't register as a compose
sequence. So maybe I should just get used to using the "TLDE" and "menu"
keys as compose keys, since there is no delay when using these.