Remove some mappings that are useless to me
These characters are either useless to me, or are (or are going to be)
produced in other ways. For example, I intend to produce the
"registered" and "trademark" using the composition key.

These "slots" then get freed up for more useful characters. I don't yet
know what to replace them with, so for now I'm just removing things.