Add feminine ordinal
I might want to use this character to conveniently mark ordinal
characters. E.g., instead of writing:

    The 1st, 2nd, 3rd…

I can write:

    The 1ª, 2ª, 3ª…

In other words, I don't have to worry about how to write the ending of
the numbers; just use the numeral plus the feminine ordinal.

This would not be intended as markup, or how my finished writing would
look like. It would solely be to clearly distinguish numerals from
ordinals. When “publishing” the document, the software could easily
convert those ordinals to:

    The 1st, 2nd, 3rd…

Or:

    The first, second, third…

Or some other style.

I am choosing the feminine ordinal because the masculine ordinal can be
mistaken for the degree sign (°).