Add four "technical" symbols
That is: I intend to use them for that. Things like meta
characters—command characters, escape characters, etc. The strength of
using non-ASCII characters is that they are less likely to occur in the
rest of the text, so there is less chance and need for them being
escaped.

These characters might seem pretty random, and they are. I tried to
choose characters in such a way that they would be hard to confuse for
each other.

Also remove the feminine ordinal, since that character can be found on
the "q" key.