Add access to levels 5–8
I looked at this file:

    /usr/share/X11/xkb/symbols/de

And noticed that the configurations that had eight levels included
either of these lines:

    key.type[Group1] = "EIGHT_LEVEL";
    key.type[Group1] = "EIGHT_LEVEL_ALPHABETIC";

That is, either of these lines came before a sequence of "key" mappings.

I chose to use the "EIGHT_LEVEL" because the difference that one and
"EIGHT_LEVEL_ALPHABETIC" seemed to be[1] that the latter included locking.
I don’t need locking right now, and don’t know how I would use it
properly. So the former seems like the best choice for me right now.

[1] To determine this I looked at the following file, which I think is
where these things are defined:

    /usr/share/X11/xkb/types/level5