Fix new (old) issue with "BKSL" and "LSGT" keys
The change that I made in the previous commit turned out to bring up an
issue that I’ve had and solved before[1]. The issue was that depressing
and holding "BKSL" key after depressing and holding the "Shift_R" key
lead to the "BKSL" key being registered as "NoSymbol". In turn, I
couldn’t access the fourth level. There also seemed to be a problem with
the "LSGT" being registered as "NoSymbol" sometimes.

I thought that adding the shifts to all 8 levels of these keys had
solved this issue for good (again, see [1]).

As luch would have it, the files:

    /usr/share/X11/xkb/symbols/level3
    /usr/share/X11/xkb/symbols/level5

had mappings for using the "BKSL" and "LSGT" keys for level 3 and level
5, respectively. So delete my own mappings of these keys to be shift
keys and include these mappings instead. Also, add "modifier_mapping"
inclusions to make sure that level 3 and 5 are mapped to “real
modifiers” (as the files in which those definitions are puts it).

[1] See commit 58d854ecc1ff7643555f37751533569671a32721