Add some combining diacritics to fifth level keys
Add five different diacritics that are “mirrored” on the middle
alphabetical row. This means that these characters can be accessed by
using the index, ring, etc. finger of each hand. The intent is to
facilitate touch typing when using these combining characters.

In addition, add six diacritics on the right side of the keyboard
only (i.e., they are not mirrored).

For keys that don’t have a fourth level defined, temporarily *pad them
out* by using "NoSymbol". I intend to find an actual character for these
later.

Since this might be hard to visualise from the diff alone, the following
shows how the middle alphabetic row and some of the keys for the lower
alphabetic row end up looking like for the fifth level (in case you
can’t see the diacritics, copy and paste the table to another
application which can display them):

┣━━━━━━━┻┱────┴┬────┴┬────┴┬────┴┬────┴┬────┴┬────┴┬────┴┬────┴┬────┴┬────┴┬────┺┓      ┃
┃        ┃ ◌̄   │ ◌̈  │  ◌̂  │  ◌́ │  ◌̀ │  ◌̀ │  ◌́  │  ◌̂ │  ◌̈ │  ◌̄│  ◌̆ │     ┃      ┃
┃Caps ⇬  ┃     │     │     │     │     │     │     │      │     │     │     │     ┃      ┃
┣━━━━━━━━┹────┬┴────┬┴────┬┴────┬┴────┬┴────┬┴────┬┴────┬┴────┬┴────┬┴────┲┷━━━━━┻━━━━━━┫
┃             │     │     │     │     │     │     │     │  ◌̧  │   ◌̇  │  ◌̄    ┃             ┃
┃Shift ⇧      │     │     │     │     │     │     │     │     │     │         ┃Shift ⇧      ┃
┣━━━━━━━┳━━━━━┷━┳━━━┷━━━┱─┴─────┴─────┴─────┴─────┴─────┴───┲━┷━━━━━╈━━━━━┻━┳━━━━━━━┳━━━┛

(This does not show the two combining diacritic keys that are one the
top alphabetic row.)

Choice of Diacritics
====================

I chose ( ̀) because it seems to be sometimes be used in Norwegian, but
I’m not quite sure. I chose ( ́) because of Spanish. I chose ( ̂) and ( ̆)
because of Esperanto, and ( ̂) is “mirrored” since it is more common
than ( ̆). I chose ( ̈) because I might end up needing it sometimes,
although I don’t use any language which uses it (to my knowledge). ( ̄)
was not chosen for a good reason. I simply couldn’t think of something
more useful to put on the two “mirrored” keys (A and Ø keys on my
keyboard). Maybe I will find some better character later. ( ̃) was not
given a mirrored key because, while it is used in Spanish, it is only
used for one letter. So it doesn’t seem as important to give a mirrored
setup as ( ́) (acute accent) which I, by the way, gave what is in my
opinion the best mirrored keys (keys F and J—operated by the two index
fingers.) The rest of the combining diacritics are not really things
that I need, so they’re mostly for the occassional use (although I could
use or set up compose key sequences for them).

Overall Plan
============

I originally wanted to use this whole level for combining characters,
like combining diacritics. But then I thought of using it for something
else as well; to have a row or more of “vacant” keys that can be used to
map characters to as I need them, and preferably dynamically (i.e., I
don’t have to recompile the keymap or restart X). For example, if I
needed a couple of special symbols in a certain context, I want to
interactively be able to assign these characters to keys on this level.
I don’t know how hard this is going to be.

In any case, by only using *some* of the keys for this level, I have
many keys left over for the above described ad hoc key assignment.