Remove "BKSL" and "CAPS" as compose keys
I don’t want to dual-purpose these keys as compose keys after all. I
explained the reason for this in a commit message[1] in the repository
that I use for my Xkb configuration files (“symbol” files). But the
short story is: dual-purposing a key with Xcape introduces a delay for
the “click and relase” key behaviour. This delay might become an issue
once I start becoming fast with using the compose key[2]. So I want to
get in the habit of just using the "MENU" and "TLDE" keys as compose
keys.

On the other hand… maybe if I never get fast enough with the compose
keys for the delay to matter, I will consider re-enabling this
dual-purpose behaviour.

[1] 17a19368195ccfa4020533c1258b1a87e358c936
[2] Hopefully that happens!