Set "CAPS" as a switch in this file
I have used "CAPS" as a third level switch for some time now. But I did
that in my `keyboard-set` script by using Xcape. Now that I don’t want
to dual-purpose it as a compose key and a shift level 3 switch, I might
as well set it in this configuration instead.