Add comment clarifying the last modifier key
The last of the shift level keys.

This setup gives a kind-of symmetric setup of the three different levels
of shift keys; two keys per level, on both sides of the keyboard. But
since the "Return" key is set somewhere else, it might be puzzling that
it seems that there is only one level 5 shift key. This comment is
meant to clarify that there are indeed two such keys.