Move necessary characters somewhere else
Move the useful characters which were on the backslash key to other
keys.

The characters that I replace are either characters that I don’t need,
or that I can access on some other key.

§. Birds-eye view

What I am doing is moving each of the two characters that I need to the
two keys to the left of the backslash key, on the third level of each
key.  Thus I can access both of these characters by pressing caps lock
and these keys at the same time, since caps lock is a level 3 shift in
my setup.  So I think this is will be only a *slight* downgrade in
efficiency for these two characters, since I need to press a shift key
in order to produce a backslash, which I didn’t need to do before.  I
don’t think this will be an annoyance in practice, since I don’t feel
that I use backslash that much.

§. Backslash character (`\`)

Move this character to the “0” key, <AE10>.  This because I can access
`braceright` on another key, which I prefer.  Move`braceright` up one
level, replacing `∅` (U+2205 EMPTY SET) which I don’t use.

§. Grave character (U+0060 GRAVE ACCENT)

Move this character to the “plus” key, <AE11>.  This because I don’t
need `quistiondown`.  Move `quistiondown` up one level.