Duplicate first five symbols to alphabetic area
Duplicate (†1) first five symbols from the numeric row to alphabetic
area.

Those five symbols are:

    !, ", @, #, $

I am going through the numeric row from left to right, including the
often-used symbols.  So I include the dollar sign, but not seldom-used
signs like the euro sign.

This will hopefully make these symbols easier type, since they’re now in
the “home row” area, as vimmers seem to call it.

This replaces five symbols that I for the most part don’t use, with the
exception of one:

    ☇

But I can eventually find another spot for this character, or use a
compose sequence to produce it.

†. 1: “Duplicate” since I am not *moving* the symbols down; the symbols
    stay on the numerical row.