Move Vi-style arrows one row down
Move Vi-style arrows (†1) one row down by moving them from the third
level of the “yiuo” keys to the fourth level of the “hjlk” keys.  This
frees up the third level of the “yiuo” keys, which I intend to use for
the backslash key and the grave key now at first.

This makes it a bit harder to produce these arrows, since you have to
hold two shift keys in order to produce fourth level keys.  But I don’t
think this will be a problem, since I seldom use these keys.  This is
also more consistent, in that the Vi-like keys are now on the Vi
movement keys instead of one row up from them.

†. 1: The Vi-style arrows were added in r#540179f7a5a1 (“Use vi-style
    layout for arrows”, 2016-10-06).