Duplicate backslash and grave keys
Duplicate backslash and grave keys on the alphabetic section by putting
them on the now-vacant third levels of the `Y` and `U` keys,
respectively.  They are now vacant since we moved the Vi-style arrows
from these keys in r#a05bd49988f9 (“Move Vi-style arrows one row down”,
2017-04-30).

These two keys are easier to reach than he `0` and `+`, which are the
other keys that backslash and grave are bound to (on the third level).