Duplicate equals and question mark to alpbahetic
Duplicate equals sign and question mark to alpbahetic section.

These two symbols are on the second level of the numeric row.  So they
should be easier to type as third level keys on the alphabetic section.

I think this concludes the duplication of often-used keys from the
numeric row to the alphabetic section.