Add Norwegian setup for keyboard_1
Try to add the standard (`no`) keymap for Norwegian keyboards.  Try to
do this by first including that keymap, and then including the `basic`
keymap in this file.  The intent is that the relevant stuff from `basic`
will override the things that I don’t want from `no`, but not things
that I don’t want and don’t impact in `basic`, like e.g. the three
keys that produce the Norwegian characters `æ`, `ø`, and `å`.