Start adding commonly used symbols
Start adding commonly used symbols to the third level of alphabetic
keys.

Add these keys:

    %, &, `, =, ?, !, ", @, #, $

These keys were already added, so they are commented out:

    |, \, ', *

They are commented out so that mappings follow each other left-to-right
as they appear on the keyboard, starting with the first row (`AD`) and
then the second row (`AC`).

The addition of twelve of these symbols corresponds to the commit that
changed `latin_p1` (†1).

†. 1: r#21747ba20f53 (“Merge branch ‘move-down-symbol-keys’”,
    2017-05-09)