Remove comments about the initially moved symbols
Remove comments about the symbols that was moved initially to make room
for using `<BKSL>` as a level 3 shift key.  They are no longer really
needed.  Also, move these mappings to be “sorte” with the rest of the
mappings, where the sorting is to put them in a left-to-right order,
with the first row and then the second row.