Duplicate `less` and `greater` symbols
Duplicate `less` and `greater` symbols, which on Norwegian keyboards are
on a key to the right of the left shift key.

This is not really necessary, since on European keyboards these symbols
seem to be accessible on the `<LSGT>` key, it seems.  At least they are
on the Norwegian variant.  But since these keys are free (on the third
level), I might as well duplicate the symbols there and make the third
level on those keys the same as the ones on the `no_p1` keymap.