Duplicate brackets and slash to the third row
Duplicate brackets and slash.  This was not done in the same effort for
`no_p1` since in `no_p1` these symbols are accessed from the Norwegian
characters (†1), so these keys are freed up for other characters.  But
since this keymapping leaves the Norwegian characters as-is, new keys
for the brackets and slash symbols need to be found.

The thought behind this is that the parentheses and slash symbols should
be easy to access, since they seem to be used the most.  For that reason
they are on the `B`, `N`, and `M` keys, since they are decently easy to
reach.

†. 1: The three different brackets are accessed on the `Ø` (left) and
    `Æ` (right) keys, spread over three levels.  Slash (`/`) is accessed
    on the first level of the `Å` key.