Adjust instructions on restarting X session
Adjust instructions on restarting X session in light of the fact that
not all mappings require.

Not all mappings require restarting the X session in order to work
properly.  So make this clear in the readme.

Also flesh out why this is required for `no_p1` to work as I intend it
to.