Add French quote marks to `Z` and `X` keys
Add French (and Norwegian) quote marks to `Z` and `X` keys on the fourth
level.

This is inspired by having the single and double quote marks on the `Q`
and `W` keys.  Those two are the first keys on the left of the first
alphabetic row, while `Z` and `X` are the two first keys on the left of
the third alphabetic row.  So hopefully this is a bit mnemonic.

Another motivation is that these two symbols are on the third level of
these keys on the `no` mapping.  So hopefully it will not be hard to get
used to if one wants to use these characters.