Rename keymapping to `twin-dexter`
Rename keymapping from `mundane_1` to `twin-dexter

I needed a project name and “Twin Dexter” is what I came up with.  What
follows is my reasoning.

“Dexter” is latin for “right-handed”.  One of the goals of my
keymappings is to be more friendly to touch typing, and to be more
symmetric (such as having modifiers on both the left and right side of
the keyboard).  Touch typing involves using both hands in a very
symmetric way.  This as opposed to using one hand more than the other.
Thus, “twin dexter” means to use both your hands as dominant hands, as
opposed to just one of them.

I initially thought of “ambidextrous”, but I wanted something less
generic.