Start making new keymap “no-p2”
Start making new keymap “no-p2”, the second “Norwegian programming”
keymapping.

Start off by mapping the `<LSGT>` (key to the right of the left shift
key) to “left shift”.  The reason for that is that I want to “expand”
the left shift key.  This kind-of does this, since now these two
adjacent keys are shift keys.  What this allows me to do is to more
easily use the `<Super_L>` and the left shift key at the same; this will
come in useful later.