Swap menu key with right Windows key instead
Swap menu key with right Windows key instead of right Alt key.

This is what I am after, since it makes the right Win key be next to the
right control key, and the menu key between the right Alt and the right
Win key (on the key labeled Win).