Change group switches for `no-p3` and `no-p4`
Replace the Return key for the Insert key for switching to the next
group, since Return is easier to reach from the home row.  Also add
Backspace for switching to the previous group, which might be useful for
setups that uses three or more layouts or groups.  Also move these
functions from the third to the fourth level, with the intent being of
making it harder to activate these by accident.  But it should still be
easy to switch groups and layouts, since I can access the fourth level
by depressing the left Shift and the Caps Lock key with a flat left
pinky finger.