Revert “Change group switches for `no-p3` and `no-p4`”
This reverts commit 508e7c803af447ec0ff1da39610165625e069d98.

This doesn’t work properly; see the issue in `issues.org` authored
today.  So go back to previous group switch which uses the Delete key,
which I don’t need that much anyway so it won’t be a big deal if it
doesn’t work as a Delete key in that setup.