Add a new keymapping for fifth level symbols
Add a new keymapping for fifth level symbols which combines an
“alpbabetic numpad” on the left side with combining characters on the
right side.

For now just copy over the right-side combining characters from other
setups.  This does not fill in all the keys on the fifth level on the
right side, or try to rearrange things to something more comfortable or
convenient.  I will expand or change this setup based on my experiences
with using combining diacritics (and I need to start using them more in
order to get more experience with them!).

What I want to use this for is to be able to easily access numbers
through the numpad on keymappings like `no-p4` (which doesn’t have
numbers on the “numeral row”), and also be able to easily access
combining characters (for diacritics).

If I use this for `no-p4`, this would replace the current fifth level
“symmetrical combining characters” setup.  This might make using
combining characters slightly less ergonomic, since I can’t access the
most common diacritics on both sides of the keyboard.  But hopefully it
will still be efficient to use.