Add control keys on the numeric row
This is mostly to add something to the “1” and “+” buttons, which have
not been assigned a modifier key or something else.  But it might turn
out to be useful in some cases, for all I know.

Add `modifier_map` part because it seemed to be needed.  Without it,
`xev` reports that holding down these keys makes it so that they get
repeated, which shouldn’t be what happens for control keys.  I should
probably do the same for the hyper keys, since I’ve seen in `xev` that
they also repeat.