Add a level 5 shift to the left side
Add a level 5 shift to the left side of the keyboard.

I want a level 5 shift here since then I can press it in conjunction
with Shift and Shift level 3.  So then I have both a level 5 shift and a
level 5 latch on the left side of the keyboard, same as the right side.