Add keyboard mapping `eo-common`
Add keyboard mapping which I intend to try to use when writing
Esperanto.

The idea is the same as the original idea I had for the first
programming layer, `no_p1`: use the keys for the Norwegian symbols,
which I don’t need when writing in Esperanto.  I only replace the `Ø`
and `Æ` symbols, since I don’t need more symbols.

§. Naming

“eo” is the language ISO 639-1 code for Esperanto.