Turn two keys into a dead key
Use two symmetric keys (on opposite sides of the keyboard, both accessed
by the pinky fingers) for a dead key which I have made a sort of partial
“Extend layer” for.

The purpose of this is to get a “latch” way to access a layer like this
with one key.  This means to have a single key that works the same way
as the latch shift keys.  I could also add this as a fifth, sixth or
seventh layer—and I intend to eventually do that—and access it by using
two shift keys, but the problem with that is that then I have to press
two keys to latch to that level instead of one.  (I might be able to map
a key to latch two keys, but I haven’t found a way to do that.)

§. Extend layer

Extend layer is made by Øystein Bech “DreymaR” Gadmar.

See: <https://github.com/DreymaR/BigBagKbdTrixXKB>
See also: <https://forum.colemak.com/topic/2014-extend-extra-extreme/>

§. dotfiles repository commit

The initial implementation of the “Extend layer” was implemented in the
dotfiles repository in commit r#6e9fdf9eff69 (“Make an “Extend layer” by
way of a dead key”, 2017-08-30).