Ignore a Git repository
This repository contains (among other things) the Extend layer which I
needed to look at when I tried to implement my own version of it in the
dotfiles repository in r#6e9fdf9eff69 (“Make an “Extend layer” by way of
a dead key”, 2017-08-30).  Downloading it and using things like Grep was
easier than finding stuff in it on GitHub.