Make `no-p5`, a slight variation on `no-p4`
Make a new layout that only changes `no-p4` by remapping the `<AD11>` to
free up the symbols used in `<AD12>`.  I don’t know what to do with
`<AD12>` yet, but I might use it for a dead key; see the journal entry
from today.  The purpose of this early version of this layout is to get
used to accessing the symbols the symbols `asciicircum` and `asciitilde`
in this new way.