Make levels 5–8 possible for `latin` as well
I intend to add a variation of the extend layer to the layouts that I
use.  It seems that I would need to add the layer to this layout as well
if I want to avoid duplicated mappings, since this layout defines a lot
of keys that the inheriting layouts do not.  Since the Extend layer is
at the seventh layer, `latin` needs to be capable of producing levels
5–8.