Add selected keys from the Extend layer
Add selected keys from the Extend layer.  Use the `extend` file in the
same directory as a guide to what keysyms to add.

I am primarily interested in navigation keys.  I am not interested in
things like keys corresponding to the scroll button of the mouse.  For
keys that I don’t need, add `NoSymbol` in order to mark them as being
vacant.

Only add keys for the three “alphabetic rows”.  Don’t add keys for the
numeric row since that one is taken by modifiers, compose keys, and so
on.  Also don’t add for the function keys.  Not for any technical
reason; I just can’t be bothered right now.

Add the extend layer on the seventh and eighth layer.  What this means
is that the same keys are mapped to the seventh and eighth layer (†1).
The point of the eighth layer is so that one can use `Shift` with the
different keys, like for example `Shift+Left`.  This is different from
the original Extend layer, which uses the fifth and sixth layer.

Also, change the formatting for mappings by putting levels 1–4 and
levels 5–8 on their own lines.

†. 1: One exception is for the key `<AB01>` which gets two different
   keys: `Undo` (seventh layer) and `Redo` (eighth layer).