Map `<AD01>` to left control on extend layers
Map the `Q` key to left control so that control can be comfortably used
with the extend layer.

Explicitly make a group in order to make an `actions` section which maps
actions to different levels.  This is only needed for the two extend
layers since we are dealing with a control key (†1).  I tried once
without the `actions` section and then this key did work as a control
key; the key repeated when held down instead of just being registered
one time.  And using it in combination with other keys like
`Right` (`Control+Right`) did not work.

This displaces `Escape`; move that to `<AC01>` instead (one row down).

†. 1: Having the `extend` file was really helpful here; I wouldn’t have
    known that I could make an `actions` section in order to set
    modifiers if I didn’t have that file to look at.