Add left alt to `<AC04>` on extend layers
Add left alt to `<AC04>` on extend layers since that key can comfortably
be pressed with the thumb while pressing the level three and level five
shift keys.  This displaces a group switch key, so move all of those
keys one row up, and in turn move the “back” and “forward” keys to make
room for those.  (I should refactor things if I want to make changes
like these in the future because this is quite labor-intensive…)