Use comma-separated layouts instead
The previous script did not work as the documentation described (pass in
all layouts to the script as (from the shell’s perspective) separate
arguments) since `-layout` expects a list of comma-separated layouts.
So instruct that in the documentation and use the first argument to the
script instead of the whole list of arguments.