Clean up the `numpad-and-combining` layout
Clean up the `numpad-and-combining` layout by mapping all levels that
are not changed to `NoSymbol`.  In this case that means all but the
fifth level.

I found out a little while ago that mapping a level to `NoSymbol` makes
it so that whatever inherited (included) layouts take effect.  Or
whatever layout that inherits this layout and defines symbols for the
`NoSymbol` levels take effect.  I had always thought that I would need
to remap all symbols for each level each time I wanted to redefine any
layer of a key, but that is not the case.

For example, say that some layout `a` defines this key:

    key <AC08> { [
        k, K, apostrophe, uparrow,
        NoSymbol, NoSymbol, Down, Down
    ] };

`numpad-and-combining` defines this key thus:

    key <AC08> { [
        NoSymbol, NoSymbol, NoSymbol, NoSymbol,
        // grave accent
        U0300, NoSymbol, NoSymbol, NoSymbol
    ] };

Another layout `b` includes both of these layout:

    include "a"
    include "numpad-and-combining"

And doesn’t redefine the key `<AC08>`.  This means that layout `b`
should get this mapping for this key:

    key <AC08> { [
        k, K, apostrophe, uparrow,
        U0300, NoSymbol, Down, Down
    ] };

Since layout `b` defines levels 1–4, 7, and 8, while
`numpad-and-combining` defines level 5.  At least this is how I think it
works.