Map `<AE12>` to `dead_belowmacron`
Map `<AE12>` to `dead_belowmacron` instead of `dead_greek`.

This breaks the previous mirroring with `<TLDE>`.  The reason for this
is that I intend to use the dead key which is mapped to
`<TLDE>`—`dead_greek`—for abbreviations (such as “btw”) and standard
data input (like my name).  I don’t feel that I get much from having a
symmetric setup with such a key, since having to type `<dead_greek> b t
w` all with the left hand is not very different from being able to type
the same thing by typing `<dead_greek>` with the right hand and the rest
of the left hand.  In other words, I don’t think there is enough of an
efficiency and ergonomics gain from being able to alternate hands with
longer sequences like that.

I do not yet know what I should use this new dead key for.