License the whole repository under MIT
At first I was unsure about what to license the whole repository under.
I’ve decided that the MIT license is probably overall the best; it is
simple and the same as the one DreymaR uses for his Big Bag of Keyboard
Tricks repository, from which I have taken the `extend` file.