Make a right control on the extend level
Make a right control on the extend level that mirrors the right control
on `<AD01>` on the extend level.

`<AD01>` (the `Q` key) is one row down and one column to the left of the
two “extend shifts” (the two shift keys you need to press to access
level 7) on the left side.  `<AD10>` is one row down and one column to
the right of the two “extend shifts” on the right side, so these two
keys mirror each other.