Map the `M` key to Alt on the extend levels
Map the `M` key to Alt on the extend levels so that I can access from
the extend levels on both the left and right side (on the left side the
`F` key is mapped to Alt).

I was wondering if I should map this to the `N` or the `M` key.  I
decided on `M` key since I think that will mean less travel in most
cases.