Use older version of Xkbcomp
Override the `src` attribute of `pkgs.xorg.xkbcomp` in order to use
version of 1.3.1 of Xkbcomp instead of 1.4.0.  This is needed because
the extend level does not work when the keyboard layouts are compiled
with version 1.4.0; see the issue titled “Extend level is not working
right now” in `issues.org`.

§. Source and acknowledgement

[Section 7.1.2] of the Nipkgs manual taught me about overriding
attributes.  I also got help on the #nixos channel to make the first Nix
expression which I used to get the older version of Xkbcomp, which I
used in my `/etc/nixos/configuration.nix`.

☇. [Section 7.1.2]:
    <https://nixos.org/nixpkgs/manual/#sec-pkg-overrideAttrs>