Swap positions of `"` and `’`
Swap positions of straight double quote (`"`) and single right
quote (`’`).

The point here is to get `’` in a better position.  This placement
centers around how it is primarily used, namely as a punctuation mark
for contractions.

I want `’` to be in a position where it feels nice and comfortable to
write common contractions.  Some of those I have in mind are:

• don’t
• won’t
• can’t
• you’ll
• I’ll
• we’ll
• he’ll
• she’ll
• they’ll
• won’t
• you’re
• they’re
• let’s

With this symbol on the third level of the `<AD03>` key (the `E` key), I
can do the following “rolls”:

• For words like “don’t”, “won’t”, etc.: place the ring finger on `E`,
  press the third shift level with the other hand, hit `E`, then hit
  `T`.  This produces the string `’t`, which is the last part of those
  words.
• For words like “you’re”: place the middle finger on `E`, press the
  third shift level with the other hand, hit `E`, then hit `R`, then hit
  `E`.  This produces the string `’re`, which is the last part of this
  word.

These combinations are “rolls”, which just means that they are sequences
of keypresses that alternate between adjacent fingers.  Notice how all
these sequences alternate between keys.  In other words, there are no
subsequences of pressing the same key in succession.

§. Comparison with the previous placement

The previous placement (on the `R` key) was good for rolling when it
came to words like “don’t” and “won’t”, but not good for words that
end in `’re`, like `they’re` and `you’re`.  The reason for that is that
you have to hit the `R` key twice in succesion; once shifted to produce
`’`, and once to produce `r`.  So it seems that having `’` on the `E`
key allows for more “rolls” when writing contractions than `R`.