Map `<AD12>` to `dead_belowmacron`
Map `<AD12>` to `dead_belowmacron` in the Norwegian layouts.

`dead_belowmacron` is mapped to three vowels with diacritics (†1) which
are used in the Norwegian language:

    ò, à, è
    ó, á, é
    ô, â, ê

These letters can be produced with this dead key with the following
keys (respectively):

    W, E, R
    S, D, F
    X, C, V

As you can see if you are using a Qwerty layout, this should make for
convenient typing from the home row.  Notice also that `<AD12>` is on
the right hand while these other keys are on the left hand, which makes
it so that each letter is produced by alternating the hands (which
should be more comfortable than pressing two keys in sucession with one
hand).

†. 1: See r#132bda3067cf (“Use `<dead_belowmacron>` for diacritic
   vowels”, 2017-11-01) in the dotfiles repository.