Move `dead_belowmacron` from `no` to `no-p4`
Move the binding of `<AD12>` to `dead_belowmacron` from `no` to
`no-p4`.

I want to keep the lower-numbered layouts more standard.  Using a custom
dead key (in the sense that it only works as intended with my own
compose mappings for it) for `<AD12>` is not standard at all.  So move
the introduction of this to a higher-numbered layout.