Map `Y` on the extend level to right control
Map `Y` on the extend level to right control so that I have an
alternative control key available on both sides of the keyboard on my
extend level.

Choose the `Y` key since it is a decently convenient key on the upper
alphabetic row.  I’ve already tried to use another key on the upper row,
`P`, but that didn’t work due to key ghosting; see
r#246c3a0fce27 (“Revert “Make a right control on the extend level””,
2017-10-26).